Q2 proposal ready to go
One question - should the estimate at the top be changed to reflect actuals, or remain there as the original estimate?

In the fund distribution section, please note the $ adjustment for the accounting error from June 1. There were 3 rows in the dOrg accounting spreadsheet that were not included in the original formula we submitted the proposal for. A correction was included in the July 1 amount.